requirements.txt
Django==1.6.7
MySQL-python==1.2.5
South>=1.0
argparse==1.2.1
dj-database-url==0.3.0
-e git://github.com/Wildhind/django-addresspicker.git@7063c934e750fe82cf9bed0317fcf8f87137c899#egg=django_addresspicker-master
django-allauth==0.18.0
django-ckeditor-updated==4.4.4
django-classy-tags==0.5.1
django-cleanup==0.1.10
django-cms==3.0.6
django-cuser==2014.9.28
django-debug-toolbar==1.2.1
django-debug-toolbar-template-timings==0.6.4
django-endless-pagination==2.0
django-form-utils==1.0.1
django-modeltranslation==0.8
django-mptt==0.6.1
django-resized==0.2.4
django-reversion==1.8.2
django-sekizai==0.7
django-widget-tweaks==1.3
djangocms-admin-style==0.2.2
djangocms-column==1.4
djangocms-file==0.0.2
djangocms-flash==0.0.3
djangocms-googlemap==0.1.0
djangocms-inherit==0.0.2
djangocms-installer==0.5.4
djangocms-link==1.5
djangocms-picture==0.0.4
djangocms-style==1.4
djangocms-teaser==0.0.2
djangocms-text-ckeditor==2.4.0
djangocms-video==0.0.2
html5lib==0.999
oauthlib==0.6.3
python-memcached==1.53
python-openid==2.2.5
pytz==2014.4
querystring-parser==1.2.0
requests==2.3.0
requests-oauthlib==0.4.1
six==1.7.3
sorl-thumbnail==11.12
sqlparse==0.1.11
wsgiref==0.1.2
awesome-slugify
django-yandex-money
pytils
